Conflicts:
	CockBlock8.1/CockBlock8.1/CockBlock8.1.Shared/ViewModel/CB_ViewModel.cs

+++ CockBlock8.1/CockBlock8.1/CockBlock8.1.Shared/ViewModel/CB_ViewModel.cs
@@ -53,7 +53,7 @@ namespace CockBlock8._1
         private void Update(object sender, object e)
         {
             _turnTimer--;
-            ((SingleGame)_currentPage).SetTime((int)(_turnTimer/TIMEFORSHOOTING)*100);
+            ((SingleGame)_currentPage).SetTime((int)(((float)_turnTimer/TIMEFORSHOOTING)*100));
 
             foreach (Player p in _players)
             {
